Factor node data out.
