+++ asdf/search.py
@@ -6,15 +6,8 @@ import inspect
 import re
 import typing
 
-from ._display import (
-    DEFAULT_MAX_COLS,
-    DEFAULT_MAX_ROWS,
-    DEFAULT_SHOW_VALUES,
-    _NodeInfo,
-    format_faint,
-    format_italic,
-    render_tree,
-)
+from ._display import DEFAULT_MAX_COLS, DEFAULT_MAX_ROWS, DEFAULT_SHOW_VALUES, format_faint, format_italic, render_tree
+from ._node_data import NodeSchemaData
 from .treeutil import get_children, is_container
 from .util import NotSet
 
@@ -335,9 +328,9 @@ class AsdfSearchResult:
             isinstance(self._node, dict)
             or isinstance(self._node, list)
             or isinstance(self._node, tuple)
-            or _NodeInfo.supports_info(self._node)
+            or NodeSchemaData.traversable(self._node)
         ):
-            if _NodeInfo.supports_info(self._node):
+            if NodeSchemaData.traversable(self._node):
                 child = self._node.__asdf_traverse__()[key]
             else:
                 child = self._node[key]
@@ -370,10 +363,10 @@ def _walk_tree_breadth_first(root_identifiers, root_node, callback):
                 isinstance(node, dict)
                 or isinstance(node, list)
                 or isinstance(node, tuple)
-                or _NodeInfo.supports_info(node)
+                or NodeSchemaData.traversable(node)
             ) and id(node) in seen:
                 continue
-            if _NodeInfo.supports_info(node):
+            if NodeSchemaData.traversable(node):
                 tnode = node.__asdf_traverse__()
             else:
                 tnode = node
